+++ packages/plugin/src/plugin/bridge/__mocks__/adapter.ts
