+++ packages/plugin/src/plugin/bridge/__tests__/adapter.test.ts
