+++ packages/plugin/src/plugin/bridge/__tests__/socket.test.ts
