+++ packages/plugin/src/plugin/bridge/adapter.ts
