+++ packages/plugin/src/plugin/bridge/pipe-path.ts
